Replace size_t with compatible type unsigned int in harness to make definition of size_t unnecessary.
We have the problem of C files that were already preprocessed with other C headers,
so our definitions of datatypes like size_t or functions, e.g., malloc, may be different and produce an error
due to multiple, different definitions.
Because of this, we do not explicity declare used standard library functions and data types and live
with the warning of gcc if the data type is not already defined in  the analyzed, already preprocessed C file.

We have to 'merge' the data because, otherwise, we would have to redefine distinctive data types defined in a C file
analyzed. Since we want to be able to handle *all* undefined functions as non-deterministic functions, this is a necessity
to evade a lot of additional effort, for now.