Replace "If gcov reports "NaN", convert it to "0"" with proper fix
This reverts commit fbb727def26552b8b38871ce3ae34ae50fad0e5f
and adds the correct fix.

The commit was born out of the misunderstanding that gcov can,
in some cases, report 'NaN' instead of proper numbers,
but the actual problem is that, if no gcov results are present,
the python value 'None' is used.
So if a result is not available, just don't display the statistics line.