Replace "--ig-options" by "--".
"--ig-options" wouldn't work
because we won't be able to differentiate between
a space in parantheses and a space outside of the same.
This would be necessary to properly parse the provided options
in the old manner.
Now, to make it easier for ourselves and the user,
everything after a "--" will be passed to the test-case generator.