Revert "Update cpatiger"
This reverts commit 32aee6225fd1e3e25fa3f1194a9dd26d3ba9eea3
and commit 749e459.

At the moment, the latest version of cpatiger
is not compatible with the way TBF works.
It seems to assume that every input variable
is only assigned once, which would require
us to know the number of input values before
the program execution - this is not feasible at the moment.